Optimize Locale Templates
Locale Template Optimization - Missing Files

Update QuestDef.cpp
Update cs_npc.cpp
Update QuestDef.cpp
Update CreatureTemplate.cpp
Update Creature.cpp
Update cs_npc.cpp
Update CreatureData.h
Update ItemTemplate.cpp
Update ObjectMgr.cpp
Update QuestDef.cpp
Update QuestDef.cpp
Update QuestDef.h
Update QuestDef.cpp
Update QuestDef.h
Config Template Locale Generate

CreatureMovement::ToString include fix

[3.3.5] Core/Entities: Template Optimizer for Locales - Quest Update

Revert "[3.3.5] Core/Entities: Template Optimizer for Locales - Quest Update"

This reverts commit 9a8094ef3b27d35ce64c89da6ff79c88f9a3803f.

Quest Template Update #2

Quest Packet Write Fix
