+++ src/server/game/Entities/Creature/Creature.cpp
@@ -53,21 +53,6 @@
 #include "WorldPacket.h"
 #include <G3D/g3dmath.h>
 
-std::string CreatureMovementData::ToString() const
-{
-    char const* const GroundStates[] = { "None", "Run", "Hover" };
-    char const* const FlightStates[] = { "None", "DisableGravity", "CanFly" };
-
-    std::ostringstream str;
-    str << std::boolalpha
-        << "Ground: " << GroundStates[AsUnderlyingType(Ground)]
-        << ", Swim: " << Swim
-        << ", Flight: " << FlightStates[AsUnderlyingType(Flight)];
-    if (Rooted)
-        str << ", Rooted";
-
-    return str.str();
-}
 
 VendorItemCount::VendorItemCount(uint32 _item, uint32 _count)
     : itemId(_item), count(_count), lastIncrementTime(GameTime::GetGameTime()) { }
@@ -104,119 +89,6 @@ VendorItem const* VendorItemData::FindItemCostPair(uint32 item_id, uint32 extend
         if (vendorItem.item == item_id && vendorItem.ExtendedCost == extendedCost)
             return &vendorItem;
     return nullptr;
-}
-
-uint32 CreatureTemplate::GetRandomValidModelId() const
-{
-    uint8 c = 0;
-    uint32 modelIDs[4];
-
-    if (Modelid1) modelIDs[c++] = Modelid1;
-    if (Modelid2) modelIDs[c++] = Modelid2;
-    if (Modelid3) modelIDs[c++] = Modelid3;
-    if (Modelid4) modelIDs[c++] = Modelid4;
-
-    return ((c>0) ? modelIDs[urand(0, c-1)] : 0);
-}
-
-uint32 CreatureTemplate::GetFirstValidModelId() const
-{
-    if (Modelid1) return Modelid1;
-    if (Modelid2) return Modelid2;
-    if (Modelid3) return Modelid3;
-    if (Modelid4) return Modelid4;
-    return 0;
-}
-
-uint32 CreatureTemplate::GetFirstInvisibleModel() const
-{
-    CreatureModelInfo const* modelInfo = sObjectMgr->GetCreatureModelInfo(Modelid1);
-    if (modelInfo && modelInfo->is_trigger)
-        return Modelid1;
-
-    modelInfo = sObjectMgr->GetCreatureModelInfo(Modelid2);
-    if (modelInfo && modelInfo->is_trigger)
-        return Modelid2;
-
-    modelInfo = sObjectMgr->GetCreatureModelInfo(Modelid3);
-    if (modelInfo && modelInfo->is_trigger)
-        return Modelid3;
-
-    modelInfo = sObjectMgr->GetCreatureModelInfo(Modelid4);
-    if (modelInfo && modelInfo->is_trigger)
-        return Modelid4;
-
-    return 11686;
-}
-
-uint32 CreatureTemplate::GetFirstVisibleModel() const
-{
-    CreatureModelInfo const* modelInfo = sObjectMgr->GetCreatureModelInfo(Modelid1);
-    if (modelInfo && !modelInfo->is_trigger)
-        return Modelid1;
-
-    modelInfo = sObjectMgr->GetCreatureModelInfo(Modelid2);
-    if (modelInfo && !modelInfo->is_trigger)
-        return Modelid2;
-
-    modelInfo = sObjectMgr->GetCreatureModelInfo(Modelid3);
-    if (modelInfo && !modelInfo->is_trigger)
-        return Modelid3;
-
-    modelInfo = sObjectMgr->GetCreatureModelInfo(Modelid4);
-    if (modelInfo && !modelInfo->is_trigger)
-        return Modelid4;
-
-    return 17519;
-}
-
-void CreatureTemplate::InitializeQueryData()
-{
-    for (uint8 loc = LOCALE_enUS; loc < TOTAL_LOCALES; ++loc)
-        QueryData[loc] = BuildQueryData(static_cast<LocaleConstant>(loc));
-}
-
-WorldPacket CreatureTemplate::BuildQueryData(LocaleConstant loc) const
-{
-    WorldPackets::Query::QueryCreatureResponse queryTemp;
-
-    std::string locName = Name, locTitle = Title;
-    if (CreatureLocale const* cl = sObjectMgr->GetCreatureLocale(Entry))
-    {
-        ObjectMgr::GetLocaleString(cl->Name, loc, locName);
-        ObjectMgr::GetLocaleString(cl->Title, loc, locTitle);
-    }
-
-    queryTemp.CreatureID = Entry;
-    queryTemp.Allow = true;
-
-    queryTemp.Stats.Name = locName;
-    queryTemp.Stats.NameAlt = locTitle;
-    queryTemp.Stats.CursorName = IconName;
-    queryTemp.Stats.Flags = type_flags;
-    queryTemp.Stats.CreatureType = type;
-    queryTemp.Stats.CreatureFamily = family;
-    queryTemp.Stats.Classification = rank;
-    memcpy(queryTemp.Stats.ProxyCreatureID, KillCredit, sizeof(uint32) * MAX_KILL_CREDIT);
-    queryTemp.Stats.CreatureDisplayID[0] = Modelid1;
-    queryTemp.Stats.CreatureDisplayID[1] = Modelid2;
-    queryTemp.Stats.CreatureDisplayID[2] = Modelid3;
-    queryTemp.Stats.CreatureDisplayID[3] = Modelid4;
-    queryTemp.Stats.HpMulti = ModHealth;
-    queryTemp.Stats.EnergyMulti = ModMana;
-    queryTemp.Stats.Leader = RacialLeader;
-
-    for (uint32 i = 0; i < MAX_CREATURE_QUEST_ITEMS; ++i)
-        queryTemp.Stats.QuestItems[i] = 0;
-
-    if (std::vector<uint32> const* items = sObjectMgr->GetCreatureQuestItemList(Entry))
-        for (uint32 i = 0; i < MAX_CREATURE_QUEST_ITEMS; ++i)
-            if (i < items->size())
-                queryTemp.Stats.QuestItems[i] = (*items)[i];
-
-    queryTemp.Stats.CreatureMovementInfoID = movementId;
-    queryTemp.Write();
-    return queryTemp.Move();
 }
 
 bool AssistDelayEvent::Execute(uint64 /*e_time*/, uint32 /*p_time*/)
