+++ src/server/game/Entities/GameObject/GameObject.cpp
@@ -42,49 +42,6 @@
 #include "World.h"
 #include <G3D/Quat.h>
 
-void GameObjectTemplate::InitializeQueryData()
-{
-    for (uint8 loc = LOCALE_enUS; loc < TOTAL_LOCALES; ++loc)
-        QueryData[loc] = BuildQueryData(static_cast<LocaleConstant>(loc));
-}
-
-WorldPacket GameObjectTemplate::BuildQueryData(LocaleConstant loc) const
-{
-    WorldPackets::Query::QueryGameObjectResponse queryTemp;
-
-    std::string locName = name;
-    std::string locIconName = IconName;
-    std::string locCastBarCaption = castBarCaption;
-
-    if (GameObjectLocale const* gameObjectLocale = sObjectMgr->GetGameObjectLocale(entry))
-    {
-        ObjectMgr::GetLocaleString(gameObjectLocale->Name, loc, locName);
-        ObjectMgr::GetLocaleString(gameObjectLocale->CastBarCaption, loc, locCastBarCaption);
-    }
-
-    queryTemp.GameObjectID = entry;
-    queryTemp.Allow = true;
-
-    queryTemp.Stats.Type = type;
-    queryTemp.Stats.DisplayID = displayId;
-    queryTemp.Stats.Name = locName;
-    queryTemp.Stats.IconName = locIconName;
-    queryTemp.Stats.CastBarCaption = locCastBarCaption;
-    queryTemp.Stats.UnkString = unk1;
-    memcpy(queryTemp.Stats.Data, raw.data, sizeof(uint32) * MAX_GAMEOBJECT_DATA);
-    queryTemp.Stats.Size = size;
-
-    for (uint32 i = 0; i < MAX_GAMEOBJECT_QUEST_ITEMS; ++i)
-        queryTemp.Stats.QuestItems[i] = 0;
-
-    if (std::vector<uint32> const* items = sObjectMgr->GetGameObjectQuestItemList(entry))
-        for (uint32 i = 0; i < MAX_GAMEOBJECT_QUEST_ITEMS; ++i)
-            if (i < items->size())
-                queryTemp.Stats.QuestItems[i] = (*items)[i];
-
-    queryTemp.Write();
-    return queryTemp.Move();
-}
 
 bool QuaternionData::isUnit() const
 {
